+++ CHANGELOG.md
@@ -14,6 +14,7 @@ request to fix it.
 ### Changed
 
 - [lustre/server_component] Fixed a bug where active connections in the client were not removed when the element is disconnected from the page.
+- [lustre/server_component] The client runtime will now attempt to reconnect if the connection is lost.
 
 ## [v5.5.1] - 2026-01-10